アプリ実行ファイルと、メモクラスのファイル名を修正しました
